scraper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omniscraper"
version = "1.0.0"
description = "A generic, config-driven scraping framework powered by Scrapling"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "K-BEL" },
]
keywords = ["scraping", "web-scraping", "scrapling", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "scrapling[fetchers]>=0.2",
    "typer[all]>=0.9",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
ai = [
    "sentence-transformers>=2.2",
    "httpx>=0.24",
    "elasticsearch[async]>=8.0",
]

[project.scripts]
omniscraper = "omniscraper.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/omniscraper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
